# tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Error": {
				"properties": {
					"error": {
						"properties": {
							"fbtrace_id": {},
							"error_subcode": {
								"type": "number",
								"format": "double"
							},
							"error_data": {
								"properties": {
									"details": {},
									"messaging_product": {
										"type": "string",
										"enum": [
											"whatsapp"
										],
										"nullable": false
									}
								},
								"required": [
									"details",
									"messaging_product"
								],
								"type": "object"
							},
							"code": {
								"type": "number",
								"format": "double",
								"description": "We recommend that you build your app's error handling around error codes rather than subcodes or HTTP response status codes."
							},
							"type": {
								"type": "string"
							},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"error_subcode",
							"error_data",
							"code",
							"type",
							"message"
						],
						"type": "object",
						"description": "Error Codes\r\n0 - AuthException - Unable to authenticate the app user.\r\n3 - Api Method - Indicates an issue involving resources or permissions.\r\n10 - Permission denied - Permission has not been granted or has been removed.\r\n190 - Expired token\r\nAPI Permissions - Permission has not been granted or has been removed.\r\n\r\nLimitation errors\r\n4 - API Limit - The app has reached its API call rate limit.\r\n80007 - WhatsApp Business account has reached the volume limit.\r\n\r\nIntegrity errors\r\n368 | 131031 - The WhatsApp Business account associated with the app has been restricted or disabled for violating a platform policy.\r\n\r\nOther errors\r\n1 - Invalid request or possible server error.\r\n2 - Temporary due to inactivity or being overloaded.\r\n100 - The request included one or more incompatible or misspelled parameters.\r\n131000 - Falha ao enviar a mensagem devido a um erro desconhecido.\r\n131005 - Permission has not been granted or has been removed.\r\n131008 - A required parameter is missing from the order.\r\n131009 - One or more parameter values are not supported, the recipient's phone number is not valid for WhatsApp, or the sender's phone number has not been added to the WhatsApp Business Platform.\r\n131016 - A service is temporarily unavailable.\r\n131021 - The recipient's and sender's phone number is the same.\r\n131026 - The recipient uses a version of WhatsApp that is not compatible with the message content.\r\n131051 - Message type is not supported.\r\n131052 | 131053 - The type of media sent by WhatsApp user is not supported.\r\n132000 - The number of variable parameter values included in the request does not match the number of variable parameters defined in the template.\r\n133004 - The server is temporarily unavailable.\r\n133010 - The phone number is not registered and connected.\r\n135000 - Failed to send message due to an unknown error with your request parameters."
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProfileInterface": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Contact": {
				"properties": {
					"wa_id": {
						"type": "string",
						"description": "ID of user or group \"xxxxxxxxxx@c.us\" for contacts, \"xxxxxxxxxx@g.us\" for groups"
					},
					"profile": {
						"$ref": "#/components/schemas/ProfileInterface"
					}
				},
				"required": [
					"wa_id",
					"profile"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FieldsContact": {
				"type": "string",
				"enum": [
					"name"
				],
				"nullable": false
			},
			"BusinessProfileInterface": {
				"properties": {
					"name": {
						"type": "string"
					},
					"about": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"messaging_product": {
						"type": "string",
						"enum": [
							"whatsapp"
						],
						"nullable": false
					},
					"profile_picture_url": {
						"type": "string"
					},
					"websites": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"vertical": {
						"type": "string"
					}
				},
				"required": [
					"profile_picture_url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FieldsBusinessContact": {
				"type": "string",
				"enum": [
					"about",
					"address",
					"description",
					"email",
					"profile_picture_url",
					"websites"
				]
			},
			"Partial_BusinessProfileInterface_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"about": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"messaging_product": {
						"type": "string",
						"enum": [
							"whatsapp"
						],
						"nullable": false
					},
					"profile_picture_url": {
						"type": "string"
					},
					"websites": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"vertical": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ReturnMedia": {
				"properties": {
					"messaging_product": {
						"type": "string",
						"enum": [
							"whatsapp"
						],
						"nullable": false
					},
					"url": {
						"type": "string"
					},
					"mime_type": {
						"type": "string"
					},
					"sha256": {
						"type": "string"
					},
					"file_size": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"messaging_product",
					"url",
					"mime_type",
					"sha256",
					"file_size",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TextMessageObject": {
				"properties": {
					"body": {
						"type": "string"
					},
					"preview_url": {
						"type": "boolean"
					}
				},
				"required": [
					"body",
					"preview_url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReactMessageObject": {
				"properties": {
					"message_id": {
						"type": "string",
						"description": "The WhatsApp Message ID (wamid) of the message on which the reaction should appear. The reaction will not be sent if:\nThe message is a reaction message\nThe message has been deleted\nIf the ID is of a message that has been deleted, the message will not be delivered."
					},
					"emoji": {
						"type": "string",
						"description": "Emoji to appear on the message."
					}
				},
				"required": [
					"message_id",
					"emoji"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MediaObject": {
				"properties": {
					"id": {
						"type": "string",
						"description": "ID of File sended for the server.\nRequired when type is audio, document, image, sticker, or video and you are not using a link."
					},
					"link": {
						"type": "string",
						"description": "Send LINK HTTPS or <base64>\nRequired when type is audio, document, image, sticker, or video and you are not using an uploaded media ID.\nThe protocol and URL of the media to be sent. Use only with HTTP/HTTPS URLs.\nDo not use this field when message type is set to text."
					},
					"caption": {
						"type": "string",
						"description": "Describes the specified image or video media."
					},
					"filename": {
						"type": "string",
						"description": "Describes the filename for the specific document. Use only with document media.\nThe extension of the filename will specify what format the document is displayed as in WhatsApp."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LocationMessageObject": {
				"properties": {
					"longitude": {
						"type": "number",
						"format": "double",
						"description": "Longitude of the location."
					},
					"latitude": {
						"type": "number",
						"format": "double",
						"description": "Latitude of the location."
					},
					"name": {
						"type": "string",
						"description": "Name of the location."
					},
					"address": {
						"type": "string",
						"description": "Address of the location. Only displayed if name is present."
					}
				},
				"required": [
					"longitude",
					"latitude"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdressesContact": {
				"properties": {
					"street": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"zip": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"country_code": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"HOME",
							"WORK"
						]
					}
				},
				"required": [
					"street"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NameContact": {
				"properties": {
					"formatted_name": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"middle_name": {
						"type": "string"
					},
					"suffix": {
						"type": "string"
					},
					"prefix": {
						"type": "string"
					}
				},
				"required": [
					"formatted_name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrgContact": {
				"properties": {
					"company": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"PhonesContact": {
				"properties": {
					"phone": {
						"type": "string",
						"description": "Automatically populated with the `wa_id` value as a formatted phone number."
					},
					"type": {
						"type": "string",
						"enum": [
							"CELL",
							"MAIN",
							"IPHONE",
							"HOME",
							"WORK"
						]
					},
					"wa_id": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UrlsContact": {
				"properties": {
					"url": {
						"type": "string",
						"description": "Not implemented yeat on WAJS"
					},
					"type": {
						"type": "string",
						"enum": [
							"HOME",
							"WORK"
						]
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ContactObject": {
				"properties": {
					"addresses": {
						"items": {
							"$ref": "#/components/schemas/AdressesContact"
						},
						"type": "array",
						"description": "Not implemented yeat on WAJS"
					},
					"birthday": {
						"type": "string",
						"format": "date-time",
						"description": "Not implemented yeat on WAJS"
					},
					"emails": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Not implemented yeat on WAJS"
					},
					"name": {
						"$ref": "#/components/schemas/NameContact"
					},
					"org": {
						"$ref": "#/components/schemas/OrgContact",
						"description": "Not implemented yeat on WAJS"
					},
					"phones": {
						"$ref": "#/components/schemas/PhonesContact"
					},
					"urls": {
						"items": {
							"$ref": "#/components/schemas/UrlsContact"
						},
						"type": "array",
						"description": "Not implemented yeat on WAJS"
					}
				},
				"required": [
					"name",
					"phones"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Buttons": {
				"properties": {
					"type": {
						"type": "string",
						"enum": [
							"reply"
						],
						"nullable": false,
						"description": "Only supported type is reply (for Reply Button)"
					},
					"title": {
						"type": "string",
						"description": "Button title. It cannot be an empty string and must be unique within the message. Emojis are supported, markdown is not. Maximum length: 20 characters."
					},
					"id": {
						"type": "string",
						"description": "Unique identifier for your button. This ID is returned in the webhook when the button is clicked by the user. Maximum length: 256 characters."
					}
				},
				"required": [
					"type",
					"title",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SectionsRows": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Sections": {
				"properties": {
					"product_items": {
						"description": "Not implemented yeat\nArray of product objects. There is a minimum of 1 product per section and a maximum of 30 products across all sections."
					},
					"rows": {
						"$ref": "#/components/schemas/SectionsRows",
						"description": "Contains a list of rows. You can have a total of 10 rows across your sections."
					},
					"title": {
						"type": "string",
						"description": "Required if the message has more than one section.\nTitle of the section."
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ActionObject": {
				"properties": {
					"button": {
						"type": "string",
						"description": "Required for List Messages.\nButton content. It cannot be an empty string and must be unique within the message. Emojis are supported, markdown is not."
					},
					"buttons": {
						"items": {
							"$ref": "#/components/schemas/Buttons"
						},
						"type": "array"
					},
					"catalog_id": {
						"type": "string",
						"description": "Unique identifier of the Facebook catalog linked to your WhatsApp Business Account."
					},
					"product_retailer_id": {
						"type": "string",
						"description": "Unique identifier of the product in a catalog."
					},
					"sections": {
						"$ref": "#/components/schemas/Sections"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"InteractiveObject": {
				"properties": {
					"action": {
						"$ref": "#/components/schemas/ActionObject"
					},
					"body": {
						"type": "string"
					},
					"footer": {
						"type": "string"
					},
					"header": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"button",
							"list",
							"product",
							"product_list"
						]
					}
				},
				"required": [
					"action",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Message": {
				"properties": {
					"messaging_product": {
						"type": "string",
						"enum": [
							"whatsapp"
						],
						"nullable": false,
						"description": "Send Messages, audio, etc"
					},
					"type": {
						"type": "string",
						"enum": [
							"text",
							"image",
							"audio",
							"document",
							"template",
							"hsm"
						],
						"description": "Required for message templates.\nThe type of message being sent."
					},
					"to": {
						"type": "string",
						"description": "ID of user or group \"xxxxxxxxxx@c.us\" for contacts, \"xxxxxxxxxx@g.us\" for groups"
					},
					"context": {
						"properties": {
							"message_id": {
								"type": "string"
							}
						},
						"required": [
							"message_id"
						],
						"type": "object",
						"description": "Reply for a message"
					},
					"recipient_type": {
						"type": "string",
						"enum": [
							"individual"
						],
						"nullable": false
					},
					"template": {
						"description": "Required for template messages.\nContains all model information."
					},
					"hsm": {
						"description": "Required for message templates.\nThe contained element for the message content. Indicates that the message is highly structured. The contained parameters provide the structure."
					},
					"text": {
						"$ref": "#/components/schemas/TextMessageObject"
					},
					"reaction": {
						"$ref": "#/components/schemas/ReactMessageObject"
					},
					"image": {
						"$ref": "#/components/schemas/MediaObject"
					},
					"location": {
						"$ref": "#/components/schemas/LocationMessageObject"
					},
					"contacts": {
						"$ref": "#/components/schemas/ContactObject"
					},
					"interactive": {
						"$ref": "#/components/schemas/InteractiveObject"
					}
				},
				"required": [
					"messaging_product",
					"to"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReturnSendedMessage": {
				"properties": {
					"messaging_product": {
						"type": "string",
						"enum": [
							"whatsapp"
						],
						"nullable": false
					},
					"contacts": {
						"properties": {
							"wa_id": {
								"type": "string"
							},
							"input": {
								"type": "string"
							}
						},
						"required": [
							"wa_id",
							"input"
						],
						"type": "object"
					},
					"messages": {
						"properties": {
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"messaging_product",
					"contacts",
					"messages"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "WPP Connect Server",
		"version": "2.0.0",
		"description": "REST API similar to WhatsApp Cloud API to connect with WhatsApp and send messages",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/{PHONE_NUMBER_ID}/{SECRET_KEY}/request_code": {
			"get": {
				"operationId": "GenerateToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"data": {},
												"token": {
													"type": "string",
													"nullable": true
												},
												"status": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"token",
												"status"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/Error"
										}
									]
								}
							}
						}
					}
				},
				"description": "Generate token for acess qr-code",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "PHONE_NUMBER_ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "SECRET_KEY",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/{PHONE_NUMBER_ID}/start": {
			"get": {
				"operationId": "StartSession",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"sucess": {
											"type": "boolean"
										}
									},
									"required": [
										"sucess"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Start the session (qrCode is send via webhook)",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "PHONE_NUMBER_ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/{PHONE_NUMBER_ID}/qr_code": {
			"get": {
				"operationId": "GetQrCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"qrCode": {
											"type": "string"
										}
									},
									"required": [
										"qrCode"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Update QR Code\nUse this call after use the /start",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "PHONE_NUMBER_ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/{PHONE_NUMBER_ID}/whatsapp_contact_profile": {
			"get": {
				"operationId": "GetContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Contact"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing contact.",
				"tags": [
					"Contacts"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "ID of user or group \"xxxxxxxxxx@c.us\" for contacts, \"xxxxxxxxxx@g.us\" for groups",
						"in": "path",
						"name": "PHONE_NUMBER_ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fields",
						"required": true,
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/components/schemas/FieldsContact"
							}
						}
					}
				]
			}
		},
		"/{PHONE_NUMBER_ID}/whatsapp_business_profile": {
			"get": {
				"operationId": "GetBusinessContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"$ref": "#/components/schemas/BusinessProfileInterface"
											},
											"type": "array"
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Within the business profile request, you can specify what you want to know from your business.",
				"tags": [
					"Contacts"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "ID of user \"xxxxxxxxxx@c.us\"",
						"in": "path",
						"name": "PHONE_NUMBER_ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fields",
						"required": true,
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/components/schemas/FieldsBusinessContact"
							}
						}
					}
				]
			},
			"post": {
				"operationId": "UpdateBusinessProfile",
				"responses": {
					"200": {
						"description": "Created"
					}
				},
				"description": "To update your profile, make a POST. In your request, you can include the parameters listed below.",
				"tags": [
					"Profile"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "ID of your user \"xxxxxxxxxx@c.us\"",
						"in": "path",
						"name": "PHONE_NUMBER_ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_BusinessProfileInterface_"
							}
						}
					}
				}
			}
		},
		"/{MEDIA_ID}": {
			"get": {
				"operationId": "GetMedia",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReturnMedia"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing media sended for server.",
				"tags": [
					"Media"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MEDIA_ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteMedia",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"sucess": {
											"type": "boolean"
										}
									},
									"required": [
										"sucess"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "To delete media, make a DELETE call to the ID of the media you want to delete.",
				"tags": [
					"Media"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MEDIA_ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/{PHONE_NUMBER_ID}/media": {
			"post": {
				"operationId": "CreateMedia",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "To upload media, make a POST call to /PHONE_NUMBER_ID/media and include the parameters listed below. All media files sent through this endpoint are encrypted and persist for 30 days, unless they are deleted earlier.",
				"tags": [
					"Media"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"payload": {
										"type": "string"
									},
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"payload",
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/{PHONE_NUMBER_ID}/messages/{MESSAGE_ID}": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Message"
								}
							}
						}
					},
					"400": {
						"description": "Incorrect request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing message.",
				"tags": [
					"Messages"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MESSAGE_ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/{PHONE_NUMBER_ID}/messages": {
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReturnSendedMessage"
								}
							}
						}
					},
					"400": {
						"description": "Incorrect request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"description": "Use the /PHONE_NUMBER_ID/messages endpoint to send text, media, contacts, location, and interactive messages, as well as message templates to your customers.",
				"tags": [
					"Messages"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "ID of your user \"xxxxxxxxxx@c.us\"",
						"in": "path",
						"name": "PHONE_NUMBER_ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Message"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"tags": [
		{
			"name": "Contacts",
			"description": "Within the business profile request, you can specify what you want to know from business, groups and contacts. "
		}
	]
}